Modularity and Welcome/Goodbye added

+++ Moderation/profanity_check.py
@@ -1,16 +1,18 @@
 from better_profanity import profanity
 import discord
-
+from Moderation import modinit
 
 async def profanity_checker (message):
-    if(profanity.contains_profanity(message.content)):
-        await message.delete()
-        embed= discord.Embed(
-            title= "Warning!! ",
-            description=f"Profanity Detected in a message sent by @{message.author.name}",
-            color= discord.Color.red()
-        )
-        await message.channel.send(embed= embed)
+    if(modinit.moderationSession):
+
+        if(profanity.contains_profanity(message.content)):
+            await message.delete()
+            embed= discord.Embed(
+                title= "Warning!! ",
+                description=f"Profanity Detected in a message sent by @{message.author.name}",
+                color= discord.Color.red()
+            )
+            await message.channel.send(embed= embed)
